+++ library/FMI4j/fmi-modeldescription/src/main/kotlin/no/mechatronics/sfi/fmi4j/modeldescription/ModelExchangeModelDescription.kt
@@ -24,6 +24,7 @@
 
 package no.mechatronics.sfi.fmi4j.modeldescription
 
+import no.mechatronics.sfi.fmi4j.modeldescription.misc.ModelExchangeData
 import java.io.Serializable
