+++ library/FMI4j/fmi-modeldescription/src/main/kotlin/no/mechatronics/sfi/fmi4j/modeldescription/misc/LogCategories.kt
@@ -39,14 +39,69 @@ data class LogCategory(
         @JacksonXmlProperty
         val description: String? = null
 
-) : Serializable
+) : Serializable {
+
+    companion object {
+        /**
+         * Log all events (during initialization and simulation).
+         */
+        const val LOG_EVENTS = "logEvents"
+
+        /**
+         * Log the solution of linear systems of equations if the solution is singular
+         * (and the tool picked one solution of the infinitely many solutions).
+         */
+        const val LOG_SINGULAR_LINEAR_SYSTEMS = "logSingularLinearSystems"
+
+        /**
+         * Log the solution of nonlinear systems of equations.
+         */
+        const val LOG_NON_LINEAR_SYSTEMS = "logNonlinearSystems"
+
+        /**
+         * Log the dynamic selection of states.
+         */
+        const val LOG_DYNAMIC_STATE_SELECTION = "logDynamicStateSelection"
+
+        /**
+         * Log messages when returning fmi2Warning status from any function.
+         */
+        const val LOG_STATUS_WARNING = "logStatusWarning"
+
+        /**
+         * Log messages when returning fmi2Discard status from any function.
+         */
+        const val LOG_STATUS_DISCARD = "logStatusDiscard"
+
+        /**
+         * Log messages when returning fmi2Error status from any function.
+         */
+        const val LOG_STATUS_ERROR = "logStatusError"
+
+        /**
+         * Log messages when returning fmi2Fatal status from any function.
+         */
+        const val LOG_STATUS_FATAL = "logStatusFatal"
+
+        /**
+         * Log messages when returning fmi2Pending status from any function.
+         */
+        const val LOG_STATUS_PENDING = "logStatusPending"
+
+        /**
+         * Log all messages
+         */
+        const val LOG_ALL = "logAll"
+    }
+
+}
 
 /**
  * @author Lars Ivar Hatledal
  */
 class LogCategories : Iterable<LogCategory>, Serializable {
 
-    val size
+    val size: Int
         get() = categories.size
 
     override fun iterator(): Iterator<LogCategory> = categories.iterator()
@@ -60,61 +115,7 @@ class LogCategories : Iterable<LogCategory>, Serializable {
     operator fun contains(category: String) = categories.map { it.name }.contains(category)
 
     override fun toString(): String {
-        return "LogCategoriesImpl(size=$size, categories=$categories)"
+        return "LogCategories(size=$size, categories=$categories)"
     }
 
 }
-
-/**
- * Log all events (during initialization and simulation).
- */
-const val LOG_EVENTS = "logEvents"
-
-/**
- * Log the solution of linear systems of equations if the solution is singular
- * (and the tool picked one solution of the infinitely many solutions).
- */
-const val LOG_SINGULAR_LINEAR_SYSTEMS = "logSingularLinearSystems"
-
-/**
- * Log the solution of nonlinear systems of equations.
- */
-const val LOG_NON_LINEAR_SYSTEMS = "logNonlinearSystems"
-
-/**
- * Log the dynamic selection of states.
- */
-const val LOG_DYNAMIC_STATE_SELECTION = "logDynamicStateSelection"
-
-/**
- * Log messages when returning fmi2Warning status from any function.
- */
-const val LOG_STATUS_WARNING = "logStatusWarning"
-
-/**
- * Log messages when returning fmi2Discard status from any function.
- */
-const val LOG_STATUS_DISCARD = "logStatusDiscard"
-
-/**
- * Log messages when returning fmi2Error status from any function.
- */
-const val LOG_STATUS_ERROR = "logStatusError"
-
-/**
- * Log messages when returning fmi2Fatal status from any function.
- */
-const val LOG_STATUS_FATAL = "logStatusFatal"
-
-/**
- * Log messages when returning fmi2Pending status from any function.
- */
-const val LOG_STATUS_PENDING = "logStatusPending"
-
-/**
- * Log all messages
- */
-const val LOG_ALL = "logAll"
-
-
-
