+++ library/FMI4j/fmi-modeldescription/src/main/kotlin/no/mechatronics/sfi/fmi4j/modeldescription/misc/fmudata.kt
@@ -22,11 +22,10 @@
  * THE SOFTWARE.
  */
 
-package no.mechatronics.sfi.fmi4j.modeldescription
+package no.mechatronics.sfi.fmi4j.modeldescription.misc
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty
-import no.mechatronics.sfi.fmi4j.modeldescription.misc.SourceFile
 import java.io.Serializable
 
 /**
